backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-mail
python-dotenv
werkzeug
pyjwt
requests
google-generativeai
flask-limiter
flask-talisman
